+++ PWBase.h
@@ -48,8 +48,8 @@
 - (NSDate *)dateValueForPreferenceKey:(NSString *)key defaultValue:(NSDate *)defaultValue;
 
 // primitive types
-- (int)intValueForPreferenceKey:(NSString *)key defaultValue:(int)defaultValue;
-- (double)doubleValueForPreferenceKey:(NSString *)key defaultValue:(double)defaultValue;
+- (NSInteger)intValueForPreferenceKey:(NSString *)key defaultValue:(NSInteger)defaultValue;
+- (CGFloat)doubleValueForPreferenceKey:(NSString *)key defaultValue:(CGFloat)defaultValue;
 - (BOOL)boolValueForPreferenceKey:(NSString *)key defaultValue:(BOOL)defaultValue;
 
 // Setter
